+++ NEWS.md
@@ -1,5 +1,7 @@
 ### acefile 0.6.1-dev
 
+-   Truncate password to 50 bytes for 1:1 compatibility with official unace.
+
 
 ### acefile 0.6.0 2017-07-15